CSV reading, working on methods.

+++ oblig5/src/main/java/no/hiof/erikvs/model/NaturalSatellite.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public abstract class NaturalSatellite extends CelestialBody {
 
     private double semiMajorAxis, eccentricity;
-    private int orbitalPeriod;
+    private double orbitalPeriod;
     private CelestialBody centralCelestialBody;
     public static final double ONE_JUPITER_MASS = 1.898E27;
     public static final double ONE_JUPITER_RADIUS = 71492;
@@ -17,7 +17,7 @@ public abstract class NaturalSatellite extends CelestialBody {
         public NaturalSatellite(){}
 
         //TODO: Removed reference to solarOrder
-        public NaturalSatellite(/*int solarOrder, */String name, double radius, double mass, double semiMajorAxis, double eccentricity, int orbitalPeriod, String pictureUrl, CelestialBody centralCelestialBody){
+        public NaturalSatellite(/*int solarOrder, */String name, double radius, double mass, double semiMajorAxis, double eccentricity, double orbitalPeriod, String pictureUrl, CelestialBody centralCelestialBody){
         super(/*solarOrder,*/ name, radius, mass, pictureUrl);
         this.semiMajorAxis = semiMajorAxis;
         this.eccentricity = eccentricity;
@@ -71,7 +71,7 @@ public abstract class NaturalSatellite extends CelestialBody {
 
     // get/set for instance variable orbitalPeriod
     @JsonProperty("orbitalPeriod")
-    public int getOrbitalPeriod() {
+    public double getOrbitalPeriod() {
         return orbitalPeriod;
     }
